+++ vite.config.js
@@ -28,35 +28,3 @@ export default defineConfig(
     },
   }),
 );
-
-// import { defineConfig } from 'vite';
-// import htmlMinifier from 'vite-plugin-html-minifier';
-// import { viteSingleFile } from "vite-plugin-singlefile"
-
-// export default defineConfig({
-//   build: {
-//     modulePreload: false,
-//     minify: 'terser',
-//     terserOptions: {
-//       compress: {
-//         arrows: true,
-//         unsafe: true,
-//         unsafe_arrows: true,
-//         unsafe_comps: true,
-//         passes: 2,
-//         ecma: 2025,
-//       },
-//       mangle: {
-//         properties: {
-//           keep_quoted: 'strict',
-//         }
-//       }
-//     }
-//   },
-//   plugins: [
-//     htmlMinifier({
-//       minify: true,
-//     }),
-//     viteSingleFile(),
-//   ],
-// });